Initial import.
Killbots is a KDE port of the classic BSD console game "robots". The 
game premise is simple. For some reason you are trapped in a room with a 
bunch of killer robots that are trying to destroy you. You must survive 
by avoiding the robots and causing them to collide with one another 
and/or the junkheaps remaining from previous collisions. The game play 
is turn based.

At any point you can teleport to a random cell on the board which may 
save you or may place you right next to a robot. In some rulesets you 
can collect energy, which can be used to teleport safely, without risk 
of landing in a dangerous cell. Some rulesets also include a faster 
breed of robot that can move twice as fast as the regular variety. Extra 
points and/or energy can be gained by "waiting out the round", meaning 
that you will be unable to move for the remainder of the round until 
either you or all the robots have been destroyed. Some rulesets allow 
the player to push junkheaps, either to form protective barricades or to 
crush robots. Crushing robots with junkheaps generally yields extra 
points or energy.

At this stage, Killbots supports SVG themes, animations, loading 
different rulesets, and
highscores.

svn path=/trunk/playground/games/killbots/; revision=783933