Removed use of "file(GLOB ...)" from all CMakeList.txt files.
GLOB causes problems because it only updates when CMake is run. If new 
files are added, "make install" will use the cached value of the glob 
and the files will be missed. This is an annoyance for rulesets and 
themes, but a major headache for source files.

CCMAIL: montel@kde.org

svn path=/trunk/KDE/kdegames/killbots/; revision=875565