A major refactoring. Took animation and sequencing code that was spread between Engine and Scene and moved it to it's own class, Coordinator.
This has greatly cleaned up Engine and Scene as they now focus souly on backend and frontend respectively. Coordinator is not quite so focused, but it now handles all code related to "who does what when and in response to what".

Ideally this massive commit will have zero effect on user visible functionality, but regressions might have snuck in.

svn path=/trunk/KDE/kdegames/killbots/; revision=941980