Backport of r1010036.
Correct yet another issue with shortcut handling. The movement actions were being stored in a separate KActionCollection so that they could be shown in the Configure Shortcuts dialog and not in the Configure Toolbars dialog. It turns out this broke a couple things. Changes to these shortcuts weren't being saved between sessions and the new shortcut scheme functionality was crashing. Now, for simplicity's sake all actions are stored in the same collection.

CCBUG:202468

svn path=/branches/KDE/4.3/kdegames/killbots/; revision=1010046