Remove Sprite::setRenderSize() as KGRC::setRenderSize() isn't virtual.
Previously Sprite::setRenderSize() was just hiding KGameRendererClient
::setRenderSize(). While this worked fine in Killbots, it's still
hackish/ugly/wrong to pretend that polymorphism is present when it
actually isn't.

In this particular case, setting the offset inside receivePixmap works
perfectly well, but we might want to consider whether there is value in
making some of the methods in KGameRenderer and friends virtual to allow
for fancier reimplementations.

svn path=/trunk/KDE/kdegames/killbots/; revision=1156149